rename io to _io. Also, export _io functions in readers
