+++ hydromt/data_catalog/data_catalog.py
@@ -33,6 +33,7 @@ from pystac import Catalog as StacCatalog
 from pystac import CatalogType, MediaType
 
 from hydromt import __version__
+from hydromt._io import _yml_from_uri_or_path
 from hydromt._utils import (
     _deep_merge,
     _partition_dictionaries,
@@ -62,7 +63,6 @@ from hydromt.data_catalog.sources import (
 )
 from hydromt.error import NoDataException, NoDataStrategy, exec_nodata_strat
 from hydromt.gis.gis_utils import _parse_geom_bbox_buffer
-from hydromt.io import _yml_from_uri_or_path
 from hydromt.plugins import PLUGINS
 from hydromt.typing import Bbox, SourceSpecDict, TimeRange
