+++ tests/components/test_config_component.py
@@ -5,8 +5,8 @@ import pytest
 from tomli_w import dump as toml_dump
 from yaml import dump as yaml_dump
 
+from hydromt._io import read_yaml
 from hydromt._utils.path import _make_config_paths_absolute, _make_config_paths_relative
-from hydromt.io import read_yaml
 from hydromt.model import Model
 from hydromt.model.components.config import ConfigComponent
 from hydromt.readers import _config_read
